+++ src/Nuterra/Maritaria/MaritariaMod.cs
@@ -1,10 +1,11 @@
-﻿using Nuterra;
+﻿using System;
+
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using UnityEngine;
+
+using Maritaria.FirstPerson;
 using Newtonsoft.Json.Linq;
+using Nuterra;
+using UnityEngine;
 
 namespace Maritaria
 {
@@ -21,7 +22,6 @@ namespace Maritaria
 
 		internal static MaritariaMod Instance { get; private set; }
 
-
 		public override void Load()
 		{
 			Instance = this;
@@ -31,15 +31,18 @@ namespace Maritaria
 			_behaviourHolder = new GameObject();
 			_behaviourHolder.AddComponent<MagnetToggleKeyBehaviour>();
 			_behaviourHolder.AddComponent<TimeOfDayKeysBehaviour>();
-			_behaviourHolder.AddComponent<FirstPersonKeyBehaviour>();
+			_behaviourHolder.AddComponent<FirstPersonController>();
 			_behaviourHolder.AddComponent<ProductionToggleKeyBehaviour>();
 			_behaviourHolder.AddComponent<Gameslynx.SuicideKeyBehaviour>();
 			UnityEngine.Object.DontDestroyOnLoad(_behaviourHolder);
 
 			Config.Load(ModConfig.Data.GetValue(Name, StringComparison.OrdinalIgnoreCase) as JObject);
 
+			BlockLoader.Register(new CockpitBlock());
+			BlockLoader.Register(new BaconBlock());
+			BlockLoader.Register(new LargeCockpitBlock());
+
 			SplashScreenHandler.Init();
 		}
-
 	}
 }