+++ src/Nuterra/Sylver/GyroRotor.cs
@@ -13,7 +13,7 @@ namespace Sylver
 	public sealed class GyroRotor : CustomBlock
 
 	{
-	
+
 		public static readonly string SpriteFile = @"Assets/Blocks/Bacon/bacon_icon.png";
 
 		public static readonly int BlockID = 8000;
@@ -25,6 +25,8 @@ namespace Sylver
 		public string Name => "Hawkeye Gyro Fan Rotor";
 
 		public string Description => "An improved version of the standard Hawkeye Rotor Fan which now have an integrated Gyro-Stabilizer";
+
+		public int Price => 1000;
 
 		public FactionSubTypes Faction => FactionSubTypes.HE;
 
@@ -40,9 +42,9 @@ namespace Sylver
 
 		{
 			Prefab = UnityEngine.GameObject.Instantiate(Singleton.Manager<ManSpawn>.inst.m_BlockPrefabs[414]);
-			
-            Prefab.SetActive(false);
-			
+
+			Prefab.SetActive(false);
+
 			GameObject.DontDestroyOnLoad(Prefab);
 
 			Prefab.name = Name;
@@ -54,7 +56,7 @@ namespace Sylver
 
 
 			Visible vis = Prefab.EnsureComponent<Visible>();
-			
+
 			SylverGyro gyro = Prefab.EnsureComponent<SylverGyro>();
 
 			vis.m_ItemType = new ItemTypeInfo(ObjectTypes.Block, BlockID);
@@ -62,4 +64,4 @@ namespace Sylver
 
 	}
 
-}	
+}